refactor: move embeds to utils.embeds
